+++ src/__tests__/userFunderPermissions.int.test.ts
@@ -4,6 +4,8 @@ import {
 	createOrUpdateFunder,
 	createOrUpdateUserFunderPermission,
 	loadSystemUser,
+	loadUserFunderPermission,
+	removeUserFunderPermission,
 } from '../database';
 import { expectTimestamp, loadTestUser } from '../test/utils';
 import {
@@ -11,6 +13,7 @@ import {
 	mockJwtWithAdminRole as authHeaderWithAdminRole,
 } from '../test/mockJwt';
 import { keycloakIdToString, Permission } from '../types';
+import { NotFoundError } from '../errors';
 
 describe('/users/funders/:funderShortcode/permissions/:permission', () => {
 	describe('PUT /', () => {
@@ -161,6 +164,200 @@ describe('/users/funders/:funderShortcode/permissions/:permission', () => {
 				permission: Permission.MANAGE,
 				userKeycloakUserId: user.keycloakUserId,
 			});
+		});
+	});
+
+	describe('DELETE /', () => {
+		it('returns 401 if the request lacks authentication', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.send()
+				.expect(401);
+		});
+
+		it('returns 401 if the authenticated user lacks permission', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeader)
+				.send()
+				.expect(401);
+		});
+
+		it('returns 400 if the userId is not a valid keycloak user ID', async () => {
+			await request(app)
+				.delete(`/users/notaguid/funders/1/permissions/${Permission.MANAGE}`)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 400 if the funder shortCode is not a valid shortCode', async () => {
+			const user = await loadTestUser();
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/not a valid short code/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 400 if the permission is not a valid permission', async () => {
+			const user = await loadTestUser();
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/ExampleInc/permissions/notAPermission`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(400);
+		});
+
+		it('returns 404 if the permission does not exist', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.MANAGE}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(404);
+		});
+
+		it('returns 404 if the permission had existed and previously been deleted', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserFunderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				funderShortCode: funder.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			await removeUserFunderPermission(
+				user.keycloakUserId,
+				funder.shortCode,
+				Permission.EDIT,
+			);
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(404);
+		});
+
+		it('deletes the user funder permission when the user has administrative credentials', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserFunderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				funderShortCode: funder.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			const permission = await loadUserFunderPermission(
+				user.keycloakUserId,
+				funder.shortCode,
+				Permission.EDIT,
+			);
+			expect(permission).toEqual({
+				funderShortCode: funder.shortCode,
+				createdAt: expectTimestamp,
+				createdBy: user.keycloakUserId,
+				permission: Permission.EDIT,
+				userKeycloakUserId: user.keycloakUserId,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeaderWithAdminRole)
+				.send()
+				.expect(204);
+			await expect(
+				loadUserFunderPermission(
+					user.keycloakUserId,
+					funder.shortCode,
+					Permission.EDIT,
+				),
+			).rejects.toThrow(NotFoundError);
+		});
+
+		it('deletes the user funder permission when the user has permission to manage the funder', async () => {
+			const user = await loadTestUser();
+			const funder = await createOrUpdateFunder({
+				shortCode: 'ExampleInc',
+				name: 'Example Inc.',
+				keycloakOrganizationId: null,
+			});
+			await createOrUpdateUserFunderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				funderShortCode: funder.shortCode,
+				permission: Permission.MANAGE,
+				createdBy: user.keycloakUserId,
+			});
+			await createOrUpdateUserFunderPermission({
+				userKeycloakUserId: user.keycloakUserId,
+				funderShortCode: funder.shortCode,
+				permission: Permission.EDIT,
+				createdBy: user.keycloakUserId,
+			});
+			const permission = await loadUserFunderPermission(
+				user.keycloakUserId,
+				funder.shortCode,
+				Permission.EDIT,
+			);
+			expect(permission).toEqual({
+				funderShortCode: funder.shortCode,
+				createdAt: expectTimestamp,
+				createdBy: user.keycloakUserId,
+				permission: Permission.EDIT,
+				userKeycloakUserId: user.keycloakUserId,
+			});
+			await request(app)
+				.delete(
+					`/users/${keycloakIdToString(user.keycloakUserId)}/funders/${funder.shortCode}/permissions/${Permission.EDIT}`,
+				)
+				.set(authHeader)
+				.send()
+				.expect(204);
+			await expect(
+				loadUserFunderPermission(
+					user.keycloakUserId,
+					funder.shortCode,
+					Permission.EDIT,
+				),
+			).rejects.toThrow(NotFoundError);
 		});
 	});
 });